rename toc to anchorNavigation
